Add support for new tools in `vimrc`
Add `coc.nvim` support.

Change `fillchars` and `listchars` to look a little fancier and not create an
ever-growing list of values.

Remove `=` from the list of `isfname` characters to fix filename
completion when filenames are up against that character such as when
assigning filenames to variables in shell scripts (e.g. `FOO=/tmp/bar`).

Stop adding `**` to `path` because while it seems handy, it can cause
problems with some built-in functionality.

Fix linting issues according to the Google Vim Script Style Guide.

Replace `ag` with `rg`.

Add new `quickfix` related mappings.

Add undo breakpoints in insert mode when deleting words or lines, and
when adding a `,` or `.`.

Add some ALE fixers for Python and Terraform to automate some code
tidiness.