Add more completions to `zsh`
Add `fzf` keybindings and completions to `zsh`.

Add `git-extras` completion to `zsh`.

Add "bookmark" support to `zsh`. `goto <bookmark>` can be used to
quickly follow symlinks in the `~/.bookmarks/` directory.

Check for the `p10k.zsh` before attempting to source it.

Add `LESS_TERMCAP` to `common/files/` to improve the look of the `less`
pager.