Change how `neovim` copies files
Use a `loop` with directories rather than `with_fileglob` to copy files
in the `neovim` role.

Start/enable `mpd` in the handler rather than as a task so it doesn't
happen until it's actually installed.

Remove `units` from the list of default packages (this comes from the
AUR and thus causes problems during a standard run of the role).

Add `vim` modelines.